requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.1.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
kaggle>=1.5.0
